06 - grpc, protobuf, pydantic, sqlalchemy, alembic/src/requirements.txt
alembic==1.14.1
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
setuptools==75.8.0
SQLAlchemy==2.0.38
